etl/main_project/etl_process/requirements.txt
backoff==2.2.1
flake8==4.0
python-dotenv==0.21.0
psycopg2==2.9.5
elasticsearch==7.17.8
requests==2.28.1
